Moved components folder, added subfolder
